The inverse of the id() function: rev_id() added
The implementation of rev_id has not yet the constant
time complexity which can be achieved by adding
sub-linear space select data structures.
The implementation was tested on all nodes of
the suffix trees of inputs ``faust.txt'' and
``zarathustra.txt''.
The current time complexity is log(n) for
inner nodes and constant for leaves.