Add 8 padding bytes at the end.
In all previous versions of the library rank data structures like
rank_support_v and rank_support_v5 make an access to possible
not allocated memory. Let b the supported bitvector, than this
possible illegal memory access happened, if b.size()%64==0
and rank(b.size()) was called. Note that, if the memory after
the bitvector was allocated the result of rank(b.size()) was correct
since the content of the memory cell was mask with a zero-mask and
did not contribute to the result.
We had to options to solve the problem:
(1) We can check for each rank call, if we are in the special
    case and to not access the memory cell.
(2) We add 8 padding bytes after the bitvector when allocating
    the memory.
We opted for the second solution, since an extra branch can
significantly worsen the runtime of the rank operation.
NOTE: The additional 8 bytes are not serialized and therefore
not counted when the size of the data structure is calculated.