Fixed alphabet size.
In the old version, the alphabet size was equal to
the alphabet size of the underlaying wavelet tree
for the character-tuples of length 2 respectively 9.
However, there may exist tuples with dummy characters.
We don't want to count those dummy characters, if
they do not also occur in the original text.