Added method to serialize and load vectors.
In util the methods
  * serialize_vector
  * load_vector
were added. The methods take a std::vector
of sdsl-objects and serialized/loads each
object. The size of the vector is not
stored. So before calling load_vector, the
vector has to be resized to its original size.

If the objects have all the same structure tree,
then the methods merges all the structure trees
by summing up the sizes of the components.