Added sampling policy classes for CSAs
It is now possible to parametrize all CSAs with a policy class
for the suffix array samples of the CSA. At this point, I have
implemented two approaches.
   * The so far used suffix array order sampling (which can lead
     to very bad runtimes if repetitions are in the text) -- see
     class `sa_order_sa_sampling` in `csa_sampling_strategy.hpp`.
   * The regular text order sampling (which needs at most \(k\)
     steps if every \(k\)-th text position is sampled. See
	 class `text_order_sa_sampling`.

Note that you can specify for text_order_sa_sampling which
bitvector and rank_support is used to determine if a
suffix is sampled or not. In both policy classes you can
also specify if a variable or fixed-width int_vector should
be used.