Restructured tests
Now the tests work like the benchmarks.
Each test is called with arguments pointing
to test input and other options. This simplified
the code for the tests a lot.
Now the Makefile is responsible for passing
the arguments. The arguments are read from
the config files.

Non existing inputs are either generated with
BitVectorGenerator.x or IntVectorGenerator.x,
or downloaded using curl and extracted using
gunzip.

For details see the README.md file.