Make cache config more flexible
This commit adds a small tweak to the way the cache config
works in sdsl. Specifically, if a specific file name is stored
in the filemap within a cache config, when calling cache_file_name,
the prestored file is used instead of the standard one. So, if
custom file locations for things like the SA or BWT are now
possible.