optimized bit select using BMI2's pdep
BMI2's pdep instruction can be used to make bit selection much faster
than the current approach.

It uses pdep in a different way than normally, setting the data as mask
and using a word as source operand which has only the i-th bit set.
As a result we get the i-th bit in data isolated. By counting the
trailing zeros (tzcnt) we get the position which is the result of
select.